Fix MacOS build issues. Partial README update.
